+++ CHANGELOG.md
@@ -7,3 +7,9 @@
 
 ## [0.0.3] - ssinn
 - Added virtualization type
+
+## [0.0.4] - ssinn
+- mv'd connections.tf to main.tf
+- variable name cleanup in resources.tf
+- cleanup of object syntax
+- terraform fmt